Fixed casting.
git-svn-id: https://svnsrv.germanlloyd.org/svn/ExtFEMIO/trunk@463 389f8db8-dd28-11e4-b689-8b400cb78fb4

+++ src/bdf/bdf_cards_cmass4.cpp
@@ -107,7 +107,7 @@ namespace dnvgl {
             void cmass4::collect_outdata(
                std::list<std::unique_ptr<format_entry> > &res) const {
 
-               if (static_cast<int>(EID) <= 0)
+               if (static_cast<long>(EID) <= 0)
                   return;
 
                res.push_back(format(cmass4::head));
